usual: Require _Noreturn
This has been standard since C11.

We keep the _NORETURN macro that is currently in use.  Using the
standard <stdnoreturn.h> header and noreturn notation has the
potential for conflicts with third-party headers that use some variant
of __attribute__((noreturn)).  So for now we keep this until we have a
better idea.  Maybe converge on pg_noreturn that is used in
PostgreSQL.